+++ packages/sdk/src/ton-connect.ts
@@ -38,7 +38,7 @@ import {
     SignMessageResponse
 } from 'src/models/methods';
 import { ConnectAdditionalRequest } from 'src/models/methods/connect/connect-additional-request';
-import { TonConnectOptions } from 'src/models/ton-connect-options';
+import { AnalyticsSettings, TonConnectOptions } from 'src/models/ton-connect-options';
 import {
     isWalletConnectionSourceJS,
     isWalletConnectionSourceWalletConnect
@@ -78,6 +78,8 @@ import { isQaModeEnabled } from './utils/qa-mode';
 import { normalizeBase64 } from './utils/base64';
 import { AnalyticsManager } from 'src/analytics/analytics-manager';
 import { BrowserEventDispatcher } from 'src/tracker/browser-event-dispatcher';
+import { EventDispatcher } from 'src/tracker/event-dispatcher';
+import { SdkActionEvent } from 'src/tracker/types';
 import { bindEventsTo } from 'src/analytics/sdk-actions-adapter';
 import { BridgePartialSession, BridgeSession } from 'src/provider/bridge/models/bridge-session';
 import { IEnvironment } from 'src/environment/models/environment.interface';
@@ -117,9 +119,9 @@ export class TonConnect implements ITonConnect {
      */
     private readonly tracker: TonConnectTracker;
 
-    private readonly walletsList = new WalletsListManager();
+    private readonly walletsList: WalletsListManager;
 
-    private readonly analytics?: AnalyticsManager;
+    private analytics?: AnalyticsManager;
 
     private readonly environment: IEnvironment;
 
@@ -174,9 +176,14 @@ export class TonConnect implements ITonConnect {
 
         this.walletsRequiredFeatures = options?.walletsRequiredFeatures;
 
+        this.environment = options?.environment ?? new DefaultEnvironment();
+
         this.walletsList = new WalletsListManager({
             walletsListSource: options?.walletsListSource,
-            cacheTTLMs: options?.walletsListCacheTTLMs
+            cacheTTLMs: options?.walletsListCacheTTLMs,
+            onDownloadDurationMeasured: (duration: number | undefined) => {
+                this.analytics?.setWalletListDownloadDuration(duration);
+            }
         });
 
         const eventDispatcher = options?.eventDispatcher ?? new BrowserEventDispatcher();
@@ -187,22 +194,7 @@ export class TonConnect implements ITonConnect {
 
         this.environment = options?.environment ?? new DefaultEnvironment();
 
-        // TODO: in production ready make flag to enable them?
-        this.analytics = new AnalyticsManager({ environment: this.environment });
-
-        const telegramUser = this.environment.getTelegramUser();
-        bindEventsTo(
-            eventDispatcher,
-            this.analytics.scoped({
-                locale: this.environment.getLocale(),
-                browser: this.environment.getBrowser(),
-                platform: this.environment.getPlatform(),
-                tg_id: telegramUser?.id,
-                tma_is_premium: telegramUser?.isPremium,
-                manifest_json_url: manifestUrl,
-                origin_url: getOriginWithPath
-            })
-        );
+        this.initAnalytics(manifestUrl, eventDispatcher, options);
 
         if (!this.dappSettings.manifestUrl) {
             throw new DappMetadataError(
@@ -210,7 +202,10 @@ export class TonConnect implements ITonConnect {
             );
         }
 
-        this.bridgeConnectionStorage = new BridgeConnectionStorage(this.dappSettings.storage);
+        this.bridgeConnectionStorage = new BridgeConnectionStorage(
+            this.dappSettings.storage,
+            this.walletsList
+        );
 
         if (!options?.disableAutoPauseConnection) {
             this.addWindowFocusAndBlurSubscriptions();
@@ -413,18 +408,20 @@ export class TonConnect implements ITonConnect {
             switch (bridgeConnectionType) {
                 case 'http':
                     provider = await BridgeProvider.fromStorage(
-                        this.dappSettings.storage,
+                        this.bridgeConnectionStorage,
                         this.analytics
                     );
                     break;
                 case 'injected':
                     provider = await InjectedProvider.fromStorage(
-                        this.dappSettings.storage,
+                        this.bridgeConnectionStorage,
                         this.analytics
                     );
                     break;
                 case 'wallet-connect':
-                    provider = await WalletConnectProvider.fromStorage(this.dappSettings.storage);
+                    provider = await WalletConnectProvider.fromStorage(
+                        this.bridgeConnectionStorage
+                    );
                     break;
                 default:
                     if (embeddedWallet) {
@@ -965,6 +962,42 @@ export class TonConnect implements ITonConnect {
         }
     }
 
+    private initAnalytics(
+        manifestUrl: string,
+        eventDispatcher: EventDispatcher<SdkActionEvent>,
+        options?: TonConnectOptions
+    ): void {
+        const analyticsSettings: AnalyticsSettings | undefined = options?.analytics;
+        const mode = analyticsSettings?.mode ?? 'telemetry';
+
+        if (mode === 'off') {
+            return;
+        }
+
+        const analytics = new AnalyticsManager({
+            environment: this.environment,
+            mode
+        });
+        this.analytics = analytics;
+
+        const telegramUser = this.environment.getTelegramUser();
+
+        const sharedAnalyticsData: Parameters<typeof analytics.scoped>[0] = {
+            browser: this.environment.getBrowser(),
+            platform: this.environment.getPlatform(),
+            manifest_json_url: manifestUrl,
+            origin_url: getOriginWithPath,
+            locale: this.environment.getLocale()
+        };
+
+        if (telegramUser) {
+            sharedAnalyticsData.tg_id = telegramUser.id;
+            sharedAnalyticsData.tma_is_premium = telegramUser.isPremium;
+        }
+
+        bindEventsTo(eventDispatcher, analytics.scoped(sharedAnalyticsData));
+    }
+
     private createProvider(
         wallet: WalletConnectionSource | Pick<WalletConnectionSourceHTTP, 'bridgeUrl'>[]
     ): Provider {
@@ -972,14 +1005,14 @@ export class TonConnect implements ITonConnect {
 
         if (!Array.isArray(wallet) && isWalletConnectionSourceJS(wallet)) {
             provider = new InjectedProvider(
-                this.dappSettings.storage,
+                this.bridgeConnectionStorage,
                 wallet.jsBridgeKey,
                 this.analytics
             );
         } else if (!Array.isArray(wallet) && isWalletConnectionSourceWalletConnect(wallet)) {
-            provider = new WalletConnectProvider(this.dappSettings.storage);
+            provider = new WalletConnectProvider(this.bridgeConnectionStorage);
         } else {
-            provider = new BridgeProvider(this.dappSettings.storage, wallet, this.analytics);
+            provider = new BridgeProvider(this.bridgeConnectionStorage, wallet, this.analytics);
         }
 
         provider.listen(this.walletEventsListener.bind(this));