+++ packages/ui-react/CHANGELOG.md
@@ -1,5 +1,56 @@
 # Changelog @tonconnect/ui-react
 
+## 2.4.0-beta.7
+
+### Patch Changes
+
+- @tonconnect/ui@2.4.0-beta.7
+
+## 2.4.0-beta.6
+
+### Patch Changes
+
+- Updated dependencies [3964cf3]
+    - @tonconnect/ui@2.4.0-beta.6
+
+## 2.4.0-beta.5
+
+### Minor Changes
+
+- 3964cf3: feat: add support for custom TonConnectUI instance in TonConnectUIProvider
+
+    Added optional `tonConnectUI` prop to `TonConnectUIProvider` to allow passing a pre-initialized
+    TonConnectUI instance. This enables:
+    - Sharing a single instance between multiple parts of the application
+    - Initializing with custom logic before passing it to the provider
+    - Using the same instance across React and non-React parts of the application
+
+    **Usage:**
+
+    ```typescript
+    import { TonConnectUIProvider } from '@tonconnect/ui-react';
+    import { TonConnectUI } from '@tonconnect/ui';
+
+    const tonConnectUI = new TonConnectUI({
+      manifestUrl: 'https://myapp.com/tonconnect-manifest.json',
+      // ... other options
+    });
+
+    function App() {
+      return (
+        <TonConnectUIProvider tonConnectUI={tonConnectUI}>
+          {/* Your app */}
+        </TonConnectUIProvider>
+      );
+    }
+    ```
+
+### Patch Changes
+
+- Updated dependencies [3964cf3]
+- Updated dependencies [3964cf3]
+    - @tonconnect/ui@2.4.0-beta.5
+
 ## 2.4.0-beta.4
 
 ### Patch Changes